refactor: Line -> RequestLine으로 이름 변경 후 내용 수정
ContentType의 메서드를 활용하여 필드에 저장
